requirements.txt
aws-cdk-lib>=2.140.0
constructs>=10.0.0

